Added some tests
But after talking to some of my "friends" online (I put friends in quotes just because I've never met them in person) I've decided not to do too many more.  Why?  Cuz this project is just for fun, not designed to be anything professional.  And writing tests is not as fun as building the language.  If there are bugs (which could happen even if all tests pass), they will be fixed like most bugs are: in "production". :)